[SCS-2659] Make CAMERA_LINK_FRAME_* variables refer to the SLAMcore transform.
Until now the CAMERA_LINK_FRAME_* would determine the position of the
realsense URDF file and not the pose of the SLAMcore pose estimation
frame.

This PR changes it. We're now using the CAMERA_LINK_FRAME_* env
variables to set the base_footprint frame relative to the SLAMcore pose
estimation frame.

As part of this PR we also hardcode the transformation of the Realsense
URDF model relative to the robot base. Before we were also setting it
based on the CAMERA_LINK_FRAME_* env variables. Continuing this though
would make it impractical and we also don't gain much from an
integrator's perspective since they would have to change the URDF model
(kobuki + mounting plates + d435) altogether.